python/pyproject.toml
[project]
name = "openqasm-parser"
authors = [
    { name = "Thierry Martinez", email = "thierry.martinez@inria.fr" },
]
version = "3.1.0.0"
dependencies = [
    "antlr4-python3-runtime==4.13.2",
]
description = "Lexer and parser generated from OpenQASM specification."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
keywords = ["openqasm", "lexer", "parser", "quantum"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Physics",
]
requires-python = ">=3.9,<3.15"

[project.urls]
Homepage = "https://github.com/qat-inria/openqasm-parser"
Documentation = "https://github.com/qat-inria/openqasm-parser/blob/main/python/README.md"
Repository = "https://github.com/qat-inria/openqasm-parser.git"
